Refactor Workspace Utils (#32)
