+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This changelog format is largely based on [Keep A Changelog](https://keepachange
 
 #### 🔨 Chores & Documentation
 
+- Refactored check api to unify across package managers
+
 ## [0.3.0] - 2022-06-27
 
 #### 🚀 New Features & Enhancements
