+++ src/check/format-output.js
@@ -1,4 +1,4 @@
-const chalk = require('chalk');
+const chalk = require("chalk");
 
 const SINGLE_INDENT = 2;
 const DOUBLE_INDENT = SINGLE_INDENT * 2;
@@ -16,8 +16,8 @@ Get a string in the format:
 const getVersionString = (version, dependencyTypeStrings) => {
   return (
     chalk.magentaBright(version.padStart(SINGLE_INDENT + version.length)) +
-    '\n' +
-    dependencyTypeStrings.join('\n')
+    "\n" +
+    dependencyTypeStrings.join("\n")
   );
 };
 
@@ -26,9 +26,9 @@ Get a string in the format:
    [type]: [...names], that is:
    direct: name1, name2
 */
-const getTypeString = ({type, names}) => {
+const getTypeString = ({ type, names }) => {
   const padded = type.padStart(DOUBLE_INDENT + type.length);
-  return chalk.yellowBright(`${padded}: `) + chalk.white(names.join(', '));
+  return chalk.yellowBright(`${padded}: `) + chalk.white(names.join(", "));
 };
 
 const format = (packages) => {
@@ -39,16 +39,16 @@ const format = (packages) => {
       const versionsStr = Object.entries(versions)
         .map(([version, depTypes]) => {
           const depTypeStrings = Object.entries(depTypes).map(([type, names]) =>
-            getTypeString({type, names})
+            getTypeString({ type, names })
           );
 
           return getVersionString(version, depTypeStrings);
         })
-        .join('\n');
+        .join("\n");
 
       return `${str}\n${versionsStr}`;
     })
-    .join('\n');
+    .join("\n");
 };
 
-module.exports = {format};
+module.exports = { format };
