+++ src/shared/index.js
@@ -1,0 +1,11 @@
+const { parseConfig } = require("./read-config");
+const { getPackageDeps } = require("./read-dependencies");
+const { getWorkspacesForPackageManager } = require("./get-workspaces");
+const { detectPackageManager } = require("./detect-package-manager");
+
+module.exports = {
+  parseConfig,
+  getPackageDeps,
+  getWorkspacesForPackageManager,
+  detectPackageManager,
+};
